requirements.txt
fastapi==0.100.0  # Adjust this version to the latest that supports pydantic>=2.7.0
uvicorn==0.30.1
pydantic>=2.7.0
pydantic-settings==2.3.4
python-dotenv==1.0.1
motor==3.2.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
email-validator==2.0.0
python-multipart==0.0.5